+++ src/next.ts
@@ -1,2 +1,2 @@
-export {createCMS} from 'alinea/adapter/next/NextCMS'
-export {createHandler} from 'alinea/adapter/next/NextHandler'
+export {createCMS} from 'alinea/adapter/next/cms'
+export {createHandler} from 'alinea/adapter/next/handler'
